Refactor to avoid duplication between disk/stream
Disk and stream components had lots of duplicated code for:
- evaluating normal velocity distributions
- evaluating chemical evolution model
- the methods `get_` + `p_x_t`, `p_v_tx`, `p_z_tx`
These have been moved into an intermediate `parametricComponent` ABC.

`parametricComponent` now inherits from `baseComponent`. The base
component can be initialized with a 5D density `p_tvxz`. Added methods
to construct all marginals from this joint.
