+++ popkinmocks/components/growing_disk.py
@@ -1,10 +1,8 @@
 import numpy as np
 from scipy import stats, special
-import matplotlib.pyplot as plt
-from matplotlib.colors import LogNorm
-from . import base
+from . import parametric
 
-class growingDisk(base.component):
+class growingDisk(parametric.parametricComponent):
     """A growing disk with age-and-space dependent velocities and enrichments
 
     The (mass-weighted) joint density of this component can be factorised as
@@ -26,14 +24,6 @@ class growingDisk(base.component):
         rotation: angle (radians) between x-axes of component and cube.
 
     """
-    def __init__(self,
-                 cube=None,
-                 center=(0,0),
-                 rotation=0.):
-        super(growingDisk, self).__init__(
-            cube=cube,
-            center=center,
-            rotation=rotation)
 
     def set_p_x_t(self,
                   q_lims=(0.5, 0.1),
@@ -115,35 +105,6 @@ class growingDisk(base.component):
                                t_dep_in=t_dep_in,
                                t_dep_out=t_dep_out)
         self.t_dep = t_dep
-
-    def set_p_z_tx(self):
-        """Set p(z|t,x) Zhu+20 given enrichment model and spatially varying t_dep
-
-        Evaluates the chemical evolution model defined in equations 3-10 of
-        Zhu et al 2020, parameterised by a spatially varying depletion
-        timescale stored by `set_t_dep`.
-
-        """
-        del_t_dep = self.t_dep[:,:,np.newaxis] - self.z_t_interp_grid['t_dep']
-        abs_del_t_dep = np.abs(del_t_dep)
-        idx_t_dep = np.argmin(abs_del_t_dep, axis=-1)
-        self.idx_t_dep = idx_t_dep
-        idx_t_dep = np.ravel(idx_t_dep)
-        z_mu = self.z_t_interp_grid['z'][idx_t_dep, :]
-        z_mu = np.reshape(z_mu, (self.cube.nx, self.cube.ny, -1))
-        z_mu = np.moveaxis(z_mu, -1, 0)
-        z_sig2 = self.ahat * z_mu**self.bhat
-        log_z_edg = self.cube.ssps.par_edges[0]
-        del_log_z = log_z_edg[1:] - log_z_edg[:-1]
-        x_xsun = 1. # i.e. assuming galaxy has hydrogen mass fraction = solar
-        lin_z_edg = 10.**log_z_edg * x_xsun
-        nrm = stats.norm(loc=z_mu, scale=z_sig2**0.5)
-        lin_z_edg = lin_z_edg[:, np.newaxis, np.newaxis, np.newaxis]
-        cdf_z_tx = nrm.cdf(lin_z_edg)
-        p_z_tx = cdf_z_tx[1:] - cdf_z_tx[:-1]
-        nrm = np.sum(p_z_tx.T * del_log_z, -1).T
-        p_z_tx /= nrm
-        self.p_z_tx = p_z_tx
 
     def set_mu_v(self,
                  q_lims=(0.5, 0.5),
@@ -248,299 +209,3 @@ class growingDisk(base.component):
                                sig_v_in_lims=sig_v_in_lims,
                                sig_v_out_lims=sig_v_out_lims)
         self.sig_v = sig
-
-    def get_p_x_t(self, density=True, light_weighted=False):
-        """Get p(x|t)
-
-        Args:
-            density (bool): whether to return probabilty density (True) or the
-            volume-element weighted probabilty (False)
-            light_weighted (bool): whether to return light-weighted (True) or
-            mass-weighted (False) quantity
-
-        Returns:
-            array
-
-        """
-        if light_weighted is False:
-            p_x_t = self.p_x_t.copy()
-            if density is False:
-                p_x_t *= (self.cube.dx * self.cube.dy)
-        else:
-            p_txz = self.get_p_txz(density=density, light_weighted=True)
-            if density is True:
-                ssps = self.cube.ssps
-                p_txz = p_txz * ssps.delta_z
-            p_tx = np.sum(p_txz, -1)
-            p_t = self.get_p_t(density=density, light_weighted=True)
-            p_x_t = (p_tx.T/p_t).T
-            p_x_t = np.einsum('txy->xyt', p_x_t)
-        return p_x_t
-
-    def get_p_z_tx(self, density=True, light_weighted=False):
-        """Get p(z|t,x)
-
-        Args:
-            density (bool): whether to return probabilty density (True) or the
-            volume-element weighted probabilty (False)
-            light_weighted (bool): whether to return light-weighted (True) or
-            mass-weighted (False) quantity
-
-        Returns:
-            array
-
-        """
-        if light_weighted is False:
-            p_z_tx = self.p_z_tx.copy()
-        else:
-            p_txz = self.get_p_txz(density=True, light_weighted=True)
-            p_tx = self.get_p_tx(density=True, light_weighted=True)
-            p_z_tx = (p_txz.T/p_tx.T).T
-            p_z_tx = np.einsum('txyz->ztxy', p_z_tx)
-        if density is False:
-            dz = self.cube.ssps.delta_z
-            na = np.newaxis
-            dz = dz[:, na, na, na]
-            p_z_tx *= dz
-        return p_z_tx
-
-    def get_p_v_tx(self, v_edg, density=True, light_weighted=False):
-        """Get p(v|t,x)
-
-        Args:
-            v_edg : array of velocity-bin edges to evaluate the quantity
-            density (bool): whether to return probabilty density (True) or the
-            volume-element weighted probabilty (False)
-            light_weighted (bool): whether to return light-weighted (True) or
-            mass-weighted (False) quantity
-
-        Returns:
-            array
-
-        """
-        na = np.newaxis
-        if light_weighted is False:
-            v_edg = v_edg[:, na, na, na]
-            norm = stats.norm(loc=self.mu_v, scale=self.sig_v)
-            p_v_tx = norm.cdf(v_edg[1:]) - norm.cdf(v_edg[:-1])
-            if density is True:
-                dv = v_edg[1:] - v_edg[:-1]
-                p_v_tx /= dv
-        else:
-            p_tvxz = self.get_p_tvxz(v_edg, density=True, light_weighted=True)
-            if density is False:
-                dv = v_edg[1:] - v_edg[:-1]
-                dv = dv[na, :, na, na, na]
-                p_tvxz = p_tvxz*dv
-            ssps = self.cube.ssps
-            p_tvx = np.sum(p_tvxz*ssps.delta_z, -1)
-            p_x_t = self.get_p_x_t(density=True, light_weighted=True)
-            p_t = self.get_p_t(density=True, light_weighted=True)
-            p_xt = p_x_t * p_t
-            p_tx = np.einsum('xyt->txy', p_xt)
-            p_tx = p_tx[:, na, :, :]
-            p_v_tx = p_tvx/p_tx
-            p_v_tx = np.einsum('tvxy->vtxy', p_v_tx)
-        return p_v_tx
-
-    def get_E_v_x(self, light_weighted=False):
-        """Get mean velocity map E[p(v|x)]
-
-        Args:
-            light_weighted (bool): whether to return light-weighted (True) or
-            mass-weighted (False) quantity
-
-        Returns:
-            array
-
-        """
-        P_t = self.get_p_t(density=False, light_weighted=light_weighted)
-        E_v_x = np.sum((P_t*self.mu_v.T).T, 0)
-        return E_v_x
-
-    def get_jth_central_moment_v_x(self, j, light_weighted=False):
-        """Get j'th central moment of velocity map E[p((v-mu_v)^j|x)]
-
-        Args:
-            light_weighted (bool): whether to return light-weighted (True) or
-            mass-weighted (False) quantity
-
-        Returns:
-            array
-
-        """
-        P_t = self.get_p_t(density=False, light_weighted=light_weighted)
-        mu = self.get_E_v_x()
-        k = np.arange(0, j+1, 2)
-        tmp1 = special.comb(j, k)
-        na = np.newaxis
-        tmp2 = (self.mu_v - mu)[na,:,:,:]**(j-k[:,na,na,na])
-        tmp3 = 1.*P_t
-        tmp4 = self.sig_v[na,:,:,:]**k[:,na,na,na]
-        tmp5 = special.factorial2(k-1)
-        muj_v_x = np.einsum('k,ktxy,t,ktxy,k->xy',
-                            special.comb(j, k),
-                            (self.mu_v - mu)[na,:,:,:]**(j-k[:,na,na,na]),
-                            P_t,
-                            self.sig_v[na,:,:,:]**k[:,na,na,na],
-                            special.factorial2(k-1))
-        return muj_v_x
-
-    def get_variance_v_x(self, light_weighted=False):
-        """Get variance velocity map
-
-        Args:
-            light_weighted (bool): whether to return light-weighted (True) or
-            mass-weighted (False) quantity
-
-        Returns:
-            array
-
-        """
-        var_v_x = self.get_jth_central_moment_v_x(
-            2,
-            light_weighted=light_weighted)
-        return var_v_x
-
-    def get_skewness_v_x(self, light_weighted=False):
-        """Get skewness of velocity map
-
-        Args:
-            light_weighted (bool): whether to return light-weighted (True) or
-            mass-weighted (False) quantity
-
-        Returns:
-            array
-
-        """
-        mu3_v_x = self.get_jth_central_moment_v_x(
-            3,
-            light_weighted=light_weighted)
-        var_v_x = self.get_jth_central_moment_v_x(
-            2,
-            light_weighted=light_weighted)
-        skewness_v_x = mu3_v_x/var_v_x**1.5
-        return skewness_v_x
-
-    def get_kurtosis_v_x(self, light_weighted=False):
-        """Get kurtosis of velocity map
-
-        Args:
-            light_weighted (bool): whether to return light-weighted (True) or
-            mass-weighted (False) quantity
-
-        Returns:
-            array
-
-        """
-        mu4_v_x = self.get_jth_central_moment_v_x(
-            4,
-            light_weighted=light_weighted)
-        var_v_x = self.get_jth_central_moment_v_x(
-            2,
-            light_weighted=light_weighted)
-        kurtosis_v_x = mu4_v_x/var_v_x**2.
-        return kurtosis_v_x
-
-    def plot_density(self,
-                     vmin=0.1,
-                     vmax=3.,
-                     show_every_nth_time=4):
-        """Plot maps of the spatial density p(x|t) at several timesteps
-
-        Plot the density between the start/end times of disk growth, which
-        depends on the CDF of p(t). Skip every N steps between these points.
-
-        Args:
-            vmin: minimum velocity for colormap
-            vmax: maximum velocity for colormap
-            show_every_nth_time (int): number of timesteps to skip between plots
-
-        """
-        t_idx_list = np.arange(*self.t_pars['idx_start_end'],
-                               show_every_nth_time)
-        t_idx_list = t_idx_list[::-1]
-        kw_imshow = {'cmap':plt.cm.gist_heat,
-                     'norm':LogNorm(vmin=vmin, vmax=vmax)}
-        for t_idx in t_idx_list:
-            t = self.cube.ssps.par_cents[1][t_idx]
-            img = self.cube.imshow(self.p_x_t[:,:,t_idx], **kw_imshow)
-            plt.gca().set_title(f't={t}')
-            plt.tight_layout()
-            plt.show()
-        return
-
-    def plot_t_dep(self):
-        """Plot map of depletion timescale used for chemical enrichment
-        """
-        kw_imshow = {'cmap':plt.cm.jet}
-        img = self.cube.imshow(self.t_dep,
-                              colorbar_label='$t_\\mathrm{dep}$',
-                              **kw_imshow)
-        plt.tight_layout()
-        plt.show()
-        return
-
-    def plot_mu_v(self,
-                  show_every_nth_time=4,
-                  vmax=None):
-        """Plot maps of the mean velocity E[p(v|t,x)] at several timesteps
-
-        Plot the map between the start/end times of disk growth, which
-        depends on the CDF of p(t). Skip every N steps between these points.
-
-        Args:
-            vmax: maximum velocity for colormap
-            show_every_nth_time (int): number of timesteps to skip between plots
-
-        """
-        if vmax is None:
-            vmax = np.max(np.abs(self.mu_v_pars['vmax_lims']))
-        cube = self.cube
-        t_idx_list = np.arange(*self.t_pars['idx_start_end'],
-                               show_every_nth_time)
-        t_idx_list = t_idx_list[::-1]
-        kw_imshow = {'vmin':-vmax, 'vmax':vmax}
-        for t_idx in t_idx_list:
-            t = self.cube.ssps.par_cents[1][t_idx]
-            self.cube.imshow(self.mu_v[t_idx,:,:], **kw_imshow)
-            plt.gca().set_title(f't={t}')
-            plt.tight_layout()
-            plt.show()
-
-    def plot_sig_v(self,
-                   show_every_nth_time=4,
-                   vmin=None,
-                   vmax=None):
-        """Plot maps of the dispersion of p(v|t,x) at several timesteps
-
-        Plot the map between the start/end times of disk growth, which
-        depends on the CDF of p(t). Skip every N steps between these points.
-
-        Args:
-            vmax: minimum velocity for colormap
-            vmax: maximum velocity for colormap
-            show_every_nth_time (int): number of timesteps to skip between plots
-
-        """
-        cube = self.cube
-        t_idx_list = np.arange(*self.t_pars['idx_start_end'],
-                               show_every_nth_time)
-        t_idx_list = t_idx_list[::-1]
-        sigs = np.concatenate((
-            self.sig_v_pars['sig_v_in_lims'],
-            self.sig_v_pars['sig_v_out_lims']
-            ))
-        if vmin is None:
-            vmin = np.min(sigs)
-        if vmax is None:
-            vmax = np.max(sigs)
-        kw_imshow = {'cmap':plt.cm.jet,
-                     'vmin':vmin,
-                     'vmax':vmax}
-        for t_idx in t_idx_list:
-            t = cube.ssps.par_cents[1][t_idx]
-            cube.imshow(self.sig_v[t_idx,:,:], **kw_imshow)
-            plt.gca().set_title(f't={t}')
-            plt.tight_layout()
-            plt.show()
