+++ popkinmocks/ifu_cube.py
@@ -488,49 +488,3 @@ class IFUCube(object):
             ax.set_xlabel('$x_1$')
             ax.set_ylabel('$x_2$')
         return img
-
-    def evaluate_ybar_from_5d_density(self, p_tvxz, v_edg):
-        """Evaluate signal from a 5D density array
-
-        Uses FFTs to perfrom the convolution between the velocity component of
-        density and the SSPs.
-
-        Args:
-            p_tvxz (array): the density array with dimensions corresponding to
-                (t,v,x1,x2,z).
-            v_edg (array): array of velocity-bin edges. Number of bins must
-                equal p_tvxz.shape[1]. One of the bins must be centered on v=0.
-
-        Returns:
-            array: the evaluated 3D signal y(w,x1,x2) where w = ln(lmd)
-
-        """
-        assert v_edg.size==p_tvxz.shape[1]+1
-        assert np.allclose(v_edg[1:]-v_edg[:-1], self.ssps.dv)
-        v = (v_edg[:-1]+v_edg[1:])/2.
-        v0_idx = np.where(v==0.)
-        assert v0_idx[0].size==1
-        v0_idx = v0_idx[0][0]
-        assert np.all(np.sort(v)==v)
-        # get quantities from ssps
-        ssps = self.ssps
-        na = np.newaxis
-        dtz = ssps.delta_t[:,na,na,na,na]*ssps.delta_z[na,na,na,na,:]
-        F_s_w_tz = np.reshape(ssps.FXw, (-1,)+ssps.par_dims)
-        F_s_w_tz = np.moveaxis(F_s_w_tz, -1, 0)
-        F_s_w_tz = F_s_w_tz[:,:,na,na,:]
-        # move v=0 to correct position for the FFT
-        p_tvxz = np.roll(p_tvxz, p_tvxz.shape[1]-v0_idx, axis=1)
-        F_p_tvxz = np.fft.rfft(p_tvxz, axis=1) * ssps.dv
-        # interpolte FFT to same shape as FFT of SSPs
-        interpolator = interpolate.interp1d(
-                    np.linspace(0, 1, F_p_tvxz.shape[1]),
-                    F_p_tvxz,
-                    axis=1,
-                    kind='cubic',
-                    bounds_error=True)
-        F_p_tvxz = interpolator(np.linspace(0, 1, F_s_w_tz.shape[1]))
-        F_y = F_s_w_tz*F_p_tvxz
-        y = np.fft.irfft(F_y, ssps.n_fft, axis=1)
-        y = np.sum(y*dtz, (0,4)) * self.dx * self.dy
-        return y
