+++ tests/test_stream_component.py
@@ -5,83 +5,83 @@ import popkinmocks as pkm
 
 @pytest.fixture
 def my_stream_ybar_trim():
-    ybar_trim = np.array(
-        [[[3.47477823e-36, 6.13394662e-27, 2.03347387e-18, 4.29193180e-11,
-         9.62488723e-11],
-        [9.65063996e-22, 1.18665143e-12, 5.97382270e-11, 1.17720860e-07,
-         1.00948149e-06],
-        [1.39645848e-16, 8.56208172e-08, 2.56543636e-06, 1.06803770e-07,
-         2.13523103e-06],
-        [5.85197087e-18, 7.77903099e-09, 3.22029967e-06, 3.13930703e-06,
-         7.34899029e-08],
-        [2.43655448e-26, 7.34493669e-16, 1.83499981e-11, 3.27514134e-11,
-         1.03533620e-14]],
+    ybar_trim = np.array([[
+        [3.58595281e-36, 6.33020057e-27, 2.09853431e-18, 4.41097781e-11,
+         9.89185428e-11],
+        [9.95940956e-22, 1.22461802e-12, 6.16495353e-11, 1.20986102e-07,
+         1.03773979e-06],
+        [1.44113779e-16, 8.83602320e-08, 2.64751679e-06, 1.09872837e-07,
+         2.19658812e-06],
+        [6.03784963e-18, 8.02469484e-09, 3.32021154e-06, 3.23370629e-06,
+         7.56592089e-08],
+        [2.51340408e-26, 7.57487921e-16, 1.89172193e-11, 3.37473421e-11,
+         1.06636959e-14]],
 
-       [[4.04040071e-36, 7.13242705e-27, 2.36448162e-18, 4.76254201e-11,
-         1.06802558e-10],
-        [1.12215658e-21, 1.37981389e-12, 6.94623825e-11, 1.30628949e-07,
-         1.12244178e-06],
-        [1.62377322e-16, 9.95581265e-08, 2.98303667e-06, 1.19571198e-07,
-         2.39047867e-06],
-        [6.78789411e-18, 9.00320422e-09, 3.70120407e-06, 3.56628733e-06,
-         8.29595787e-08],
-        [2.81859479e-26, 8.47169160e-16, 2.10600904e-11, 3.73566044e-11,
-         1.17483378e-14]],
+       [[4.14736554e-36, 7.32124963e-27, 2.42707848e-18, 4.89963590e-11,
+         1.09876963e-10],
+        [1.15186435e-21, 1.41634284e-12, 7.13013170e-11, 1.34389216e-07,
+         1.15457569e-06],
+        [1.66676069e-16, 1.02193810e-07, 3.06200904e-06, 1.22941379e-07,
+         2.45785563e-06],
+        [6.96835623e-18, 9.24336629e-09, 3.80095247e-06, 3.66416077e-06,
+         8.52610880e-08],
+        [2.89383651e-26, 8.69881872e-16, 2.16288809e-11, 3.83751665e-11,
+         1.20713605e-14]],
 
-       [[3.58999746e-36, 6.33734050e-27, 2.10090128e-18, 4.35388138e-11,
-         9.76381248e-11],
-        [9.97064292e-22, 1.22599929e-12, 6.17190707e-11, 1.19420038e-07,
-         1.02163084e-06],
-        [1.44276327e-16, 8.84598948e-08, 2.65050297e-06, 1.07652384e-07,
-         2.15219661e-06],
-        [6.00850674e-18, 7.95844186e-09, 3.26981771e-06, 3.16639571e-06,
-         7.40198118e-08],
-        [2.49103344e-26, 7.48278303e-16, 1.86108986e-11, 3.30896971e-11,
-         1.04389430e-14]],
+       [[3.69023890e-36, 6.51429441e-27, 2.15956354e-18, 4.48110908e-11,
+         1.00491274e-10],
+        [1.02490475e-21, 1.26023216e-12, 6.34424168e-11, 1.22909692e-07,
+         1.05145746e-06],
+        [1.48304872e-16, 9.09299096e-08, 2.72451144e-06, 1.10783304e-07,
+         2.21479025e-06],
+        [6.17744701e-18, 8.18310983e-09, 3.36295807e-06, 3.25755263e-06,
+         7.61608297e-08],
+        [2.56141056e-26, 7.69505286e-16, 1.91418268e-11, 3.40391471e-11,
+         1.07397321e-14]],
 
-       [[3.78744372e-36, 6.68588786e-27, 2.21644874e-18, 4.97776197e-11,
-         1.11628982e-10],
-        [1.05190183e-21, 1.29342802e-12, 6.51135576e-11, 1.36532090e-07,
-         1.16492593e-06],
-        [1.52211380e-16, 9.33251002e-08, 2.79627797e-06, 1.21536792e-07,
-         2.42977501e-06],
-        [6.38608076e-18, 8.50079691e-09, 3.53886016e-06, 3.49412784e-06,
-         8.25047614e-08],
-        [2.66353252e-26, 8.04691140e-16, 2.01924721e-11, 3.62759761e-11,
-         1.15401640e-14]],
+       [[3.88457784e-36, 6.85735651e-27, 2.27329257e-18, 5.10110490e-11,
+         1.14395013e-10],
+        [1.07887927e-21, 1.32659973e-12, 6.67834830e-11, 1.39915190e-07,
+         1.19384384e-06],
+        [1.56115047e-16, 9.57185487e-08, 2.86799230e-06, 1.24572600e-07,
+         2.49046717e-06],
+        [6.54982019e-18, 8.71856963e-09, 3.62915656e-06, 3.58251253e-06,
+         8.45807390e-08],
+        [2.73175086e-26, 8.25268613e-16, 2.07071999e-11, 3.71965211e-11,
+         1.18318078e-14]],
 
-       [[3.96946737e-36, 7.00721006e-27, 2.32297075e-18, 5.06784618e-11,
-         1.13649168e-10],
-        [1.10245598e-21, 1.35558986e-12, 6.82428999e-11, 1.39002956e-07,
-         1.18751146e-06],
-        [1.59526623e-16, 9.78102826e-08, 2.93066643e-06, 1.24487770e-07,
-         2.48877124e-06],
-        [6.69250109e-18, 8.90161800e-09, 3.69194304e-06, 3.61706631e-06,
-         8.50205935e-08],
-        [2.78851551e-26, 8.41215056e-16, 2.10473226e-11, 3.76581055e-11,
-         1.19366847e-14]],
+       [[4.06475549e-36, 7.17542000e-27, 2.37873427e-18, 5.18963341e-11,
+         1.16380312e-10],
+        [1.12892073e-21, 1.38813116e-12, 6.98810888e-11, 1.42343387e-07,
+         1.21604963e-06],
+        [1.63356102e-16, 1.00158244e-07, 3.00101786e-06, 1.27479535e-07,
+         2.54858289e-06],
+        [6.85315285e-18, 9.11532331e-09, 3.78061335e-06, 3.70397828e-06,
+         8.70637458e-08],
+        [2.85546266e-26, 8.61414665e-16, 2.15528588e-11, 3.85628547e-11,
+         1.22235109e-14]],
 
-       [[3.92736266e-36, 6.93288358e-27, 2.29833066e-18, 5.04963296e-11,
-         1.13240727e-10],
-        [1.09076207e-21, 1.34121093e-12, 6.75190377e-11, 1.38503397e-07,
-         1.18350928e-06],
-        [1.57834501e-16, 9.67727949e-08, 2.89958043e-06, 1.24097062e-07,
-         2.48096017e-06],
-        [6.63122442e-18, 8.82832745e-09, 3.66919702e-06, 3.60222255e-06,
-         8.47245691e-08],
-        [2.76606402e-26, 8.35243756e-16, 2.09245911e-11, 3.74822596e-11,
-         1.18892889e-14]],
+       [[4.01871211e-36, 7.09414066e-27, 2.35178924e-18, 5.16670477e-11,
+         1.15866125e-10],
+        [1.11613292e-21, 1.37240715e-12, 6.90895131e-11, 1.41714490e-07,
+         1.21094419e-06],
+        [1.61505691e-16, 9.90237051e-08, 2.96702393e-06, 1.26973206e-07,
+         2.53846031e-06],
+        [6.78530508e-18, 9.03334999e-09, 3.75432358e-06, 3.68573037e-06,
+         8.66883040e-08],
+        [2.83029481e-26, 8.54629773e-16, 2.14099820e-11, 3.83513544e-11,
+         1.21648980e-14]],
 
-       [[3.68495827e-36, 6.50497265e-27, 2.15647327e-18, 4.61727791e-11,
-         1.03544933e-10],
-        [1.02343814e-21, 1.25842880e-12, 6.33516326e-11, 1.26644586e-07,
-         1.08235411e-06],
-        [1.48092652e-16, 9.07997915e-08, 2.72061274e-06, 1.13668530e-07,
-         2.27247196e-06],
-        [6.20034000e-18, 8.23626899e-09, 3.40453527e-06, 3.31952155e-06,
-         7.78441843e-08],
-        [2.57940202e-26, 7.76988144e-16, 1.93966547e-11, 3.46153033e-11,
-         1.09500542e-14]]])
+       [[3.77494524e-36, 6.66382459e-27, 2.20913451e-18, 4.73349287e-11,
+         1.06151116e-10],
+        [1.04843058e-21, 1.28915973e-12, 6.48986845e-11, 1.29832177e-07,
+         1.10957300e-06],
+        [1.51709086e-16, 9.30171296e-08, 2.78705032e-06, 1.16517652e-07,
+         2.32943188e-06],
+        [6.35224117e-18, 8.43846764e-09, 3.48855791e-06, 3.40205484e-06,
+         7.97867004e-08],
+        [2.64275255e-26, 7.96115422e-16, 1.98758234e-11, 3.54738371e-11,
+         1.12224863e-14]]])
 
     return ybar_trim
 
@@ -129,15 +129,15 @@ def test_component_normalisation(my_stream_component):
     assert np.allclose(np.sum(a, 0), 1.)
     a = stream.get_p_z_tx(density=True, light_weighted=True)
     assert np.allclose(np.sum(a*ssps.delta_z[:,na,na,na], 0), 1.)
-    # check p_z_t
-    a = stream.get_p_z_t(density=False, light_weighted=False)
-    assert np.allclose(np.sum(a, 0), 1.)
-    a = stream.get_p_z_t(density=True, light_weighted=False)
-    assert np.allclose(np.sum(a*ssps.delta_z[:,na], 0), 1.)
-    a = stream.get_p_z_t(density=False, light_weighted=True)
-    assert np.allclose(np.sum(a, 0), 1.)
-    a = stream.get_p_z_t(density=True, light_weighted=True)
-    assert np.allclose(np.sum(a*ssps.delta_z[:,na], 0), 1.)
+    # # check p_z_t
+    # a = stream.get_p_z_t(density=False, light_weighted=False)
+    # assert np.allclose(np.sum(a, 0), 1.)
+    # a = stream.get_p_z_t(density=True, light_weighted=False)
+    # assert np.allclose(np.sum(a*ssps.delta_z[:,na], 0), 1.)
+    # a = stream.get_p_z_t(density=False, light_weighted=True)
+    # assert np.allclose(np.sum(a, 0), 1.)
+    # a = stream.get_p_z_t(density=True, light_weighted=True)
+    # assert np.allclose(np.sum(a*ssps.delta_z[:,na], 0), 1.)
     # check p_txz
     a = stream.get_p_txz(density=False, light_weighted=False)
     assert np.isclose(np.sum(a), 1)
@@ -238,9 +238,6 @@ def test_component_ybar(
     ssps, cube, stream = my_stream_component
     ybar_trim = stream.ybar[::150, ::2, ::2]
     assert np.allclose(ybar_trim, my_stream_ybar_trim)
-
-
-
 
 
 # end
